% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdensity_d1_d2.R
\name{d_density_d1_log}
\alias{d_density_d1_log}
\title{function to return log density for D[1]
out=d_density_d1_log(y=(1:4000)/2000,D=c(0.7,5.4),S=c(7,8),1,10,hyper,1)
plot(exp(out-max(out)),type='l')}
\usage{
d_density_d1_log(y, D, S, dimIndex, N, hyper = NULL, bLog = 1)
}
\description{
function to return log density for D[1]
out=d_density_d1_log(y=(1:4000)/2000,D=c(0.7,5.4),S=c(7,8),1,10,hyper,1)
plot(exp(out-max(out)),type='l')
}
