% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finiteMixtureML.R
\name{finiteMixtureML}
\alias{finiteMixtureML}
\title{parametric mixture modeling of Matrix Langevin with fixed clusters}
\usage{
finiteMixtureML(data, nc, max_iter = 2, run_id = 1)
}
\arguments{
\item{data}{input-data coming from Stiefel array of 2D matrices data[,,i] is in V_{3,2}}

\item{nc}{numer of clusters to fit the model}

\item{max_iter}{total number of MCMC ierations}
}
\description{
parametric mixture modeling of Matrix Langevin with fixed clusters
}
