% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{print_debug}
\alias{print_debug}
\title{It is a debug print function}
\usage{
print_debug(str, debug_flag)
}
\arguments{
\item{str}{= input string}

\item{debug_flag}{= boolean flag to print or not}
}
\description{
It is a debug print function
}
